Make Pomodoro Clock unbreakable by BACK.
When a Pomodoro clock starts, cannot stop it except it counts down to 0 or cancelled. BACK activity, cannot interrupt or stop the clock. And after re-enter activity, previous clock should resume instead of starting a new clock.
So, BACK would move whole task to back.
This is not the best solution. We will use a service to do the clock.